fix: delete static route should consider dualstack

+++ pkg/controller/pod.go
@@ -879,7 +879,7 @@ func (c *Controller) handleUpdatePod(key string) error {
 						klog.Errorf("failed to add static route, %v", err)
 						return err
 					}
-				} else {
+				} else if c.config.EnableEipSnat {
 					if err := c.ovnLegacyClient.DeleteStaticRoute(podIP, c.config.ClusterRouter); err != nil {
 						return err
 					}
